Moved truffle-config
